split did creation
